// src/utils/open-im-sdk-wasm/types/entity.d.ts
export declare type WSEvent<T = unknown> = {
    event: CbEvents;
    data: T;
    errCode: number;
    errMsg: string;
    operationID: string;
};
// ...
export declare type RtcInvite = {
    inviterUserID: string;
    inviteeUserIDList: string[];
    customData?: string;
    groupID: string;
    roomID: string;
    timeout: number;
    mediaType: string;
    sessionType: number;
    platformID: number;
    initiateTime?: number;
    busyLineUserIDList?: string[];
};